perf: optimize Lighthouse mobile performance with early mobile bail-out and WebP images
Move mobile UA detection to an inline script in index.html that runs before
the module script. Extract the full desktop provider tree into a lazy-loaded
DesktopApp.tsx so mobile devices skip downloading ~3MB of desktop JS entirely.
Convert background PNG (997KB→73KB) and rocket GIF to WebP. Add preconnect
hints for API origins. Lighthouse mobile score: 60 → 94.
